InfoKiosk 0.18.0
The database has been updated, the export of current data from the database is located in the rooms.js file, the import and export scripts to the database have been updated. From now on, the News and FAQ pages are loaded from files, and the state of the pages is pulled into the files every 4 hours.